requirements.txt
requests~=2.28.2
python-dotenv~=1.0.0
apscheduler~=3.10.1
beautifulsoup4~=4.11.2
openai~=0.27.2